agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "eth-keys>=0.6.0",
    "fastapi==0.111.1",
    "numpy>=2.1.2",
    "openai>=1.52.2",
    "pydantic==2.8.2",
    "python-dotenv>=1.0.1",
    "requests==2.31.0",
    "sqlalchemy==2.0.31",
    "tweepy>=4.14.0",
    "twitter-api-client>=0.10.22",
    "uvicorn==0.30.3",
    "web3>=7.4.0",
]
